+++ internal/consumergroups/consumer-group-operation.go
@@ -43,6 +43,7 @@ type partitionOffset struct {
 type consumerGroupMember struct {
 	ClientHost         string           `json:"clientHost" yaml:"clientHost"`
 	ClientID           string           `json:"clientId" yaml:"clientId"`
+	GroupInstanceID    string           `json:"groupInstanceId" yaml:"groupInstanceId"`
 	AssignedPartitions []topicPartition `json:"assignedPartitions" yaml:"assignedPartitions"`
 }
 
@@ -138,7 +139,7 @@ func (operation *ConsumerGroupOperation) DescribeConsumerGroup(flags DescribeCon
 
 			assignedPartitions := filterAssignedPartitions(memberAssignment.Topics, topicPartitions)
 
-			consumerGroupDescription.Members = addMember(consumerGroupDescription.Members, member.ClientHost, member.ClientId, assignedPartitions)
+			consumerGroupDescription.Members = addMember(consumerGroupDescription.Members, member.ClientHost, member.ClientId, member.GroupInstanceId, assignedPartitions)
 		}
 
 		sort.Slice(consumerGroupDescription.Members, func(i, j int) bool {
@@ -173,14 +174,14 @@ func (operation *ConsumerGroupOperation) DescribeConsumerGroup(flags DescribeCon
 			consumerGroupDescription.Members = nil
 		} else if flags.OutputFormat == "wide" || flags.OutputFormat == "" {
 			tableWriter := output.CreateTableWriter()
-			if err := tableWriter.WriteHeader("CLIENT_HOST", "CLIENT_ID", "TOPIC", "ASSIGNED_PARTITIONS"); err != nil {
+			if err := tableWriter.WriteHeader("CLIENT_HOST", "CLIENT_ID", "GROUP_INSTANCE_ID", "TOPIC", "ASSIGNED_PARTITIONS"); err != nil {
 				return err
 			}
 
 			for _, m := range consumerGroupDescription.Members {
 				for _, topic := range m.AssignedPartitions {
 					partitions := strings.Trim(strings.Join(strings.Fields(fmt.Sprint(topic.Partitions)), ","), "[]")
-					if err := tableWriter.Write(m.ClientHost, m.ClientID, topic.Name, partitions); err != nil {
+					if err := tableWriter.Write(m.ClientHost, m.ClientID, m.GroupInstanceID, topic.Name, partitions); err != nil {
 						return err
 					}
 				}
@@ -226,7 +227,7 @@ func filterAssignedPartitions(assignedPartitions map[string][]int32, topicPartit
 	return result
 }
 
-func addMember(members []consumerGroupMember, clientHost string, clientID string, assignedPartitions map[string][]int32) []consumerGroupMember {
+func addMember(members []consumerGroupMember, clientHost string, clientID string, groupInstanceID *string, assignedPartitions map[string][]int32) []consumerGroupMember {
 
 	topicPartitionList := make([]topicPartition, 0)
 
@@ -247,7 +248,13 @@ func addMember(members []consumerGroupMember, clientHost string, clientID string
 	if len(assignedPartitions) == 0 {
 		return members
 	}
-	member := consumerGroupMember{ClientHost: clientHost, ClientID: clientID, AssignedPartitions: topicPartitionList}
+
+	groupInstanceIDString := ""
+	if groupInstanceID != nil {
+		groupInstanceIDString = *groupInstanceID
+	}
+
+	member := consumerGroupMember{ClientHost: clientHost, ClientID: clientID, GroupInstanceID: groupInstanceIDString, AssignedPartitions: topicPartitionList}
 	return append(members, member)
 }